Release 0.9.0. See release notes or expand full commit message.
[Breaking]
Removed ISharedEventStreamHandler and the ListeningEventStreamHandlers property on it. This pattern is not recommended for wiring update events, manage at the point of object instantiation instead.
Removed TargetId from EventStream{T}, but not EventStreamEntry{T}. This allows creating "synthetic" event streams that combine the domains of multiple event stream handlers under a single context.
Removed IEventStreamHandler.AllEventStreamEntries. Similar to 'ListeningEventStreamHandlers', if this is needed it should be managed at the point of object instantiation instead.

[Improvements]
Updated Microsoft.Bcl.AsyncInterfaces to 9.0.0.
Added net8.0 and net9.0 as TargetFrameworks.